added room use case

+++ internal/room/usecase/room_usecase.go
@@ -1,26 +1,24 @@
 package usecase
 
 import (
-	"github.com/babon21/hotel-management/domain"
+	"github.com/babon21/hotel-management/internal/domain"
 )
 
 // RoomUsecase represent the room's usecases
 type RoomUsecase interface {
 	GetList(sortField SortField, sortOrder SortOrder) ([]domain.Room, error)
-	Add(room *domain.Room) (uint64, error)
-	Delete(roomId int64) error
+	Add(room *domain.Room) (string, error)
+	Delete(roomId string) error
 }
 
 type roomUsecase struct {
 	roomRepo RoomRepository
-	//contextTimeout time.Duration
 }
 
 // NewRoomUsecase will create new an roomUsecase object representation of domain.RoomUsecase interface
 func NewRoomUsecase(roomRepository RoomRepository) RoomUsecase {
 	return &roomUsecase{
 		roomRepo: roomRepository,
-		//contextTimeout: timeout,
 	}
 }
 
@@ -28,12 +26,13 @@ func (useCase *roomUsecase) GetList(sortField SortField, sortOrder SortOrder) ([
 	return useCase.roomRepo.GetList(sortField, sortOrder)
 }
 
-func (useCase *roomUsecase) Add(room *domain.Room) (uint64, error) {
-	roomId, err := useCase.roomRepo.Save(room)
-	return roomId, err
+func (useCase *roomUsecase) Add(room *domain.Room) (string, error) {
+	return useCase.roomRepo.Save(room)
 }
 
-func (useCase *roomUsecase) Delete(roomId int64) error {
-	// TODO check exists room with specified roomId
+func (useCase *roomUsecase) Delete(roomId string) error {
+	if !useCase.roomRepo.CheckExistence(roomId) {
+		return domain.ErrNotFound
+	}
 	return useCase.roomRepo.Remove(roomId)
 }
